feat(ui): Implement main window for ESP32 LED Controller with parameter menu and serial console
- Added `mainwindow.py` to create the main application window with a parameter menu and serial console.
- Integrated `DeviceSelector` for selecting serial ports.
- Implemented functionality to toggle the visibility of the serial console and send parameters.

refactor(parameter_menu): Enhance menu structure and parameter mapping

- Updated the menu tree to include "Animation Type" and its subcategories.
- Added new parameters for "Point Control" in the parameter map.
- Adjusted layout settings for better UI experience.

fix(serial_console): Improve serial console functionality with echo and verbose options

- Added checkboxes for echo and verbose modes in the serial console.
- Enhanced command handling to support echo and verbose settings.

feat(meshnet): Define LED rigs for improved animation control

- Introduced a vector of `LEDRig` structures to manage multiple LED configurations.
- Each rig can have different animations and settings.

refactor(parameter_manager): Streamline parameter handling and listener responses

- Changed parameter listener signature to return a boolean indicating success.
- Simplified parameter checking functions for better readability.

fix(serial): Add echo functionality to serial communication

- Implemented echo feature to allow feedback of sent commands.
- Adjusted serial reading logic to handle echo commands.

feat(shared): Add utility functions for parameter validation

- Introduced functions to check if a parameter is of type bool, int, or float.
- Added string splitting utility for menu handling.

refactor(strip_state): Enhance LED strip state management

- Updated `StripState` to handle new parameters for hue and current LED.
- Improved animation handling and parameter response logic.

chore: Clean up commented code and improve logging

- Removed unnecessary commented-out code for better clarity.
- Enhanced logging for parameter handling and state changes.

+++ src/animations.cpp
@@ -206,7 +206,7 @@ void RainbowAnimation::update(StripState *strip)
 {
     float scrollSpeed = getFloat(PARAM_SCROLL_SPEED);
     float repeat = getFloat(PARAM_RAINBOW_REPEAT);
-    float offset = getFloat(PARAM_RAINBOW_OFFSET);
+    float offset = getInt(PARAM_RAINBOW_OFFSET);
     int brightness = getInt(PARAM_BRIGHTNESS);
 
     scrollPos += scrollSpeed;
